model_inference/requirements.txt
fastapi[standard]==0.115.12
pydantic==2.11.3
uvicorn==0.32.1
openai==1.77.0
google-auth==2.40.3
google-genai==1.29.0